refactor: HomeActivity에서 Screen 분리
