+++ feature-ui-home/src/main/kotlin/team/duckie/app/android/feature/ui/home/screen/DuckieHomeScreen.kt
@@ -1,0 +1,153 @@
+/*
+ * Designed and developed by Duckie Team, 2022
+ *
+ * Licensed under the MIT.
+ * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
+ */
+
+package team.duckie.app.android.feature.ui.home.screen
+
+import androidx.compose.animation.Crossfade
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.Layout
+import androidx.compose.ui.layout.MeasurePolicy
+import androidx.compose.ui.layout.layoutId
+import team.duckie.app.android.feature.ui.home.component.DuckTestBottomNavigation
+import team.duckie.app.android.feature.ui.home.constants.BottomNavigationStep
+import team.duckie.app.android.feature.ui.home.screen.guide.HomeGuideScreen
+import team.duckie.app.android.feature.ui.home.screen.mypage.MyPageScreen
+import team.duckie.app.android.feature.ui.home.screen.ranking.RankingScreen
+import team.duckie.app.android.feature.ui.home.screen.ranking.viewmodel.RankingViewModel
+import team.duckie.app.android.feature.ui.home.screen.search.SearchMainScreen
+import team.duckie.app.android.feature.ui.home.viewmodel.HomeViewModel
+import team.duckie.app.android.feature.ui.home.viewmodel.state.HomeState
+import team.duckie.app.android.util.compose.asLoose
+import team.duckie.app.android.util.compose.getPlaceable
+import team.duckie.app.android.util.compose.systemBarPaddings
+import team.duckie.quackquack.ui.component.QuackDivider
+
+private const val HomeCrossFacadeLayoutId = "HomeCrossFacade"
+private const val HomeBottomNavigationDividerLayoutId = "HomeBottomNavigationDivider"
+private const val HomeBottomNavigationViewLayoutId = "HomeBottomNavigation"
+private const val HomeGuideLayoutId = "HomeGuideLayoutId"
+
+private fun getHomeMeasurePolicy(
+    guideVisible: Boolean,
+) = MeasurePolicy { measurables, constraints ->
+    val looseConstraints = constraints.asLoose()
+    val extraLooseConstraints = constraints.asLoose(height = true)
+
+    val homeBottomNavigationDividerPlaceable =
+        measurables.getPlaceable(HomeBottomNavigationDividerLayoutId, looseConstraints)
+
+    val homeBottomNavigationPlaceable =
+        measurables.getPlaceable(HomeBottomNavigationViewLayoutId, looseConstraints)
+
+    val homeBottomNavigationHeight = homeBottomNavigationPlaceable.height
+
+    val homeBottomNavigationDividerHeight = homeBottomNavigationDividerPlaceable.height
+
+    val homeCrossFadeConstraints = constraints.copy(
+        minHeight = 0,
+        maxHeight = constraints.maxHeight - homeBottomNavigationHeight,
+    )
+    val homeCrossFacadePlaceable =
+        measurables.getPlaceable(HomeCrossFacadeLayoutId, homeCrossFadeConstraints)
+
+    val homeGuidePlaceable =
+        measurables.getPlaceable(HomeGuideLayoutId, extraLooseConstraints)
+
+    layout(
+        width = constraints.maxWidth,
+        height = constraints.maxHeight,
+    ) {
+        homeCrossFacadePlaceable.place(
+            x = 0,
+            y = 0,
+        )
+
+        homeBottomNavigationDividerPlaceable.place(
+            x = 0,
+            y = constraints.maxHeight - homeBottomNavigationHeight - homeBottomNavigationDividerHeight,
+        )
+
+        homeBottomNavigationPlaceable.place(
+            x = 0,
+            y = constraints.maxHeight - homeBottomNavigationHeight,
+        )
+        if (guideVisible) {
+            homeGuidePlaceable.place(
+                x = 0,
+                y = 0,
+            )
+        }
+    }
+}
+
+@Composable
+internal fun HomeScreen(
+    homeViewModel: HomeViewModel,
+    rankingViewModel: RankingViewModel,
+    state: HomeState,
+) {
+    Layout(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(color = MaterialTheme.colors.background)
+            .padding(systemBarPaddings),
+        content = {
+            HomeGuideScreen(
+                modifier = Modifier.layoutId(HomeGuideLayoutId),
+                onClose = { homeViewModel.updateGuideVisible(visible = false) },
+            )
+            // TODO(limsaehyun): 추후에 QuackCrossfade 로 교체 필요
+            Crossfade(
+                modifier = Modifier.layoutId(HomeCrossFacadeLayoutId),
+                targetState = state.bottomNavigationStep,
+            ) { page ->
+                when (page) {
+                    BottomNavigationStep.HomeScreen -> DuckieHomeScreen()
+                    BottomNavigationStep.SearchScreen -> SearchMainScreen()
+                    BottomNavigationStep.RankingScreen -> RankingScreen(
+                        viewModel = rankingViewModel,
+                        navigateToCreateProblem = {
+                            homeViewModel.navigateToCreateProblem()
+                        },
+                        navigateToDetail = { examId ->
+                            homeViewModel.navigateToHomeDetail(examId)
+                        },
+                    )
+
+                    BottomNavigationStep.MyPageScreen -> MyPageScreen(
+                        navigateToMyPage = {
+                            homeViewModel.navigateToNotification()
+                        },
+                        navigateToSettingPage = {
+                            homeViewModel.navigateToSetting()
+                        },
+                    )
+                }
+            }
+            QuackDivider(
+                modifier = Modifier.layoutId(HomeBottomNavigationDividerLayoutId),
+            )
+            DuckTestBottomNavigation(
+                modifier = Modifier.layoutId(HomeBottomNavigationViewLayoutId),
+                selectedIndex = state.bottomNavigationStep.index,
+                onClick = { index ->
+                    homeViewModel.navigationPage(
+                        BottomNavigationStep.toStep(index),
+                    )
+                },
+            )
+        },
+        measurePolicy = getHomeMeasurePolicy(
+            guideVisible = state.guideVisible
+        )
+    )
+}
