+++ feature-ui-home/src/main/kotlin/team/duckie/app/android/feature/ui/home/screen/HomeActivity.kt
@@ -19,10 +19,12 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Divider
 import androidx.compose.material.MaterialTheme
+import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.Layout
+import androidx.compose.ui.layout.MeasurePolicy
 import androidx.compose.ui.layout.layoutId
 import com.google.firebase.crashlytics.ktx.crashlytics
 import com.google.firebase.ktx.Firebase
@@ -40,12 +42,14 @@ import team.duckie.app.android.feature.ui.home.screen.ranking.viewmodel.RankingV
 import team.duckie.app.android.feature.ui.home.screen.search.SearchMainScreen
 import team.duckie.app.android.feature.ui.home.viewmodel.HomeViewModel
 import team.duckie.app.android.feature.ui.home.viewmodel.sideeffect.HomeSideEffect
+import team.duckie.app.android.feature.ui.home.viewmodel.state.HomeState
 import team.duckie.app.android.feature.ui.search.screen.SearchActivity
 import team.duckie.app.android.feature.ui.setting.screen.SettingActivity
 import team.duckie.app.android.navigator.feature.createproblem.CreateProblemNavigator
 import team.duckie.app.android.navigator.feature.detail.DetailNavigator
 import team.duckie.app.android.navigator.feature.notification.NotificationNavigator
 import team.duckie.app.android.util.compose.asLoose
+import team.duckie.app.android.util.compose.getPlaceable
 import team.duckie.app.android.util.compose.systemBarPaddings
 import team.duckie.app.android.util.kotlin.AllowMagicNumber
 import team.duckie.app.android.util.kotlin.fastFirstOrNull
@@ -54,13 +58,9 @@ import team.duckie.app.android.util.ui.BaseActivity
 import team.duckie.app.android.util.ui.const.Extras
 import team.duckie.app.android.util.ui.startActivityWithAnimation
 import team.duckie.quackquack.ui.color.QuackColor
+import team.duckie.quackquack.ui.component.QuackDivider
 import team.duckie.quackquack.ui.theme.QuackTheme
 import javax.inject.Inject
-
-private const val HomeCrossFacadeLayoutId = "HomeCrossFacade"
-private const val HomeBottomNavigationDividerLayoutId = "HomeBottomNavigationDivider"
-private const val HomeBottomNavigationViewLayoutId = "HomeBottomNavigation"
-private const val HomeGuideLayoutId = "HomeGuideLayoutId"
 
 @AllowMagicNumber("앱 종료 시간에 대해서 매직 넘버 처리")
 @AndroidEntryPoint
@@ -105,121 +105,11 @@ class HomeActivity : BaseActivity() {
             }
 
             QuackTheme {
-                Layout(
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .background(color = MaterialTheme.colors.background)
-                        .padding(systemBarPaddings),
-                    content = {
-                        HomeGuideScreen(
-                            modifier = Modifier.layoutId(HomeGuideLayoutId),
-                            onClose = { homeViewModel.updateGuideVisible(visible = false) },
-                        )
-                        // TODO(limsaehyun): 추후에 QuackCrossfade 로 교체 필요
-                        Crossfade(
-                            modifier = Modifier.layoutId(HomeCrossFacadeLayoutId),
-                            targetState = state.bottomNavigationStep,
-                        ) { page ->
-                            when (page) {
-                                BottomNavigationStep.HomeScreen -> DuckieHomeScreen()
-                                BottomNavigationStep.SearchScreen -> SearchMainScreen()
-                                BottomNavigationStep.RankingScreen -> RankingScreen(
-                                    viewModel = rankingViewModel,
-                                    navigateToCreateProblem = {
-                                        createProblemNavigator.navigateFrom(activity = this)
-                                    },
-                                    navigateToDetail = { examId ->
-                                        detailNavigator.navigateFrom(
-                                            activity = this,
-                                            intentBuilder = {
-                                                putExtra(Extras.ExamId, examId)
-                                            },
-                                        )
-                                    },
-                                )
-
-                                BottomNavigationStep.MyPageScreen -> MyPageScreen(
-                                    navigateToMyPage = {
-                                        notificationNavigator.navigateFrom(activity = this)
-                                    },
-                                    navigateToSettingPage = {
-                                        homeViewModel.navigateToSetting()
-                                    },
-                                )
-                            }
-                        }
-                        // TODO(limsaehyun): 추후에 QuackDivider 로 교체 필요
-                        Divider(
-                            modifier = Modifier.layoutId(HomeBottomNavigationDividerLayoutId),
-                            color = QuackColor.Gray3.composeColor,
-                        )
-                        DuckTestBottomNavigation(
-                            modifier = Modifier.layoutId(HomeBottomNavigationViewLayoutId),
-                            selectedIndex = state.bottomNavigationStep.index,
-                            onClick = { index ->
-                                homeViewModel.navigationPage(
-                                    BottomNavigationStep.toStep(index),
-                                )
-                            },
-                        )
-                    },
-                ) { measurables, constraints ->
-                    val looseConstraints = constraints.asLoose()
-                    val extraLooseConstraints = constraints.asLoose(height = true)
-
-                    val homeBottomNavigationDividerPlaceable =
-                        measurables.fastFirstOrNull { measurable ->
-                            measurable.layoutId == HomeBottomNavigationDividerLayoutId
-                        }?.measure(looseConstraints) ?: npe()
-
-                    val homeBottomNavigationPlaceable =
-                        measurables.fastFirstOrNull { measurable ->
-                            measurable.layoutId == HomeBottomNavigationViewLayoutId
-                        }?.measure(looseConstraints) ?: npe()
-
-                    val homeBottomNavigationHeight = homeBottomNavigationPlaceable.height
-
-                    val homeBottomNavigationDividerHeight =
-                        homeBottomNavigationDividerPlaceable.height
-
-                    val homeCrossFadeConstraints = constraints.copy(
-                        minHeight = 0,
-                        maxHeight = constraints.maxHeight - homeBottomNavigationHeight,
-                    )
-                    val homeCrossFacadePlaceable = measurables.fastFirstOrNull { measurable ->
-                        measurable.layoutId == HomeCrossFacadeLayoutId
-                    }?.measure(homeCrossFadeConstraints) ?: npe()
-
-                    val homeGuidePlaceable = measurables.fastFirstOrNull { measurable ->
-                        measurable.layoutId == HomeGuideLayoutId
-                    }?.measure(extraLooseConstraints) ?: npe()
-
-                    layout(
-                        width = constraints.maxWidth,
-                        height = constraints.maxHeight,
-                    ) {
-                        homeCrossFacadePlaceable.place(
-                            x = 0,
-                            y = 0,
-                        )
-
-                        homeBottomNavigationDividerPlaceable.place(
-                            x = 0,
-                            y = constraints.maxHeight - homeBottomNavigationHeight - homeBottomNavigationDividerHeight,
-                        )
-
-                        homeBottomNavigationPlaceable.place(
-                            x = 0,
-                            y = constraints.maxHeight - homeBottomNavigationHeight,
-                        )
-                        if (state.guideVisible) {
-                            homeGuidePlaceable.place(
-                                x = 0,
-                                y = 0,
-                            )
-                        }
-                    }
-                }
+                HomeScreen(
+                    homeViewModel = homeViewModel,
+                    rankingViewModel = rankingViewModel,
+                    state = state,
+                )
             }
         }
     }
@@ -264,9 +154,14 @@ class HomeActivity : BaseActivity() {
                 startActivityWithAnimation<SettingActivity>()
             }
 
+            is HomeSideEffect.NavigateToNotification -> {
+                notificationNavigator.navigateFrom(activity = this)
+            }
+
             HomeSideEffect.ClickRankingRetry -> {
                 rankingViewModel.clickRetryRanking()
             }
         }
     }
 }
+
