+++ feature-ui-home/src/main/kotlin/team/duckie/app/android/feature/ui/home/viewmodel/HomeViewModel.kt
@@ -338,18 +338,19 @@ internal class HomeViewModel @Inject constructor(
         )
     }
 
-    /** 문제 만들기 화면으로 이동한다.*/
+    /** 문제 만들기 화면으로 이동한다 */
     fun navigateToCreateProblem() = intent {
-        postSideEffect(
-            HomeSideEffect.NavigateToCreateProblem,
-        )
+        postSideEffect(HomeSideEffect.NavigateToCreateProblem)
     }
 
     /** 설정 화면으로 이동한다 */
     fun navigateToSetting() = intent {
-        postSideEffect(
-            HomeSideEffect.NavigateToSetting,
-        )
+        postSideEffect(HomeSideEffect.NavigateToSetting)
+    }
+
+    /** 알림 화면으로 이동한다 */
+    fun navigateToNotification() = intent {
+        postSideEffect(HomeSideEffect.NavigateToNotification)
     }
 
     fun updateGuideVisible(visible: Boolean) = intent {
