+++ feature-ui-home/src/main/kotlin/team/duckie/app/android/feature/ui/home/viewmodel/sideeffect/HomeSideEffect.kt
@@ -37,5 +37,7 @@ internal sealed class HomeSideEffect {
      */
     object NavigateToCreateProblem : HomeSideEffect()
 
+    object NavigateToNotification : HomeSideEffect()
+
     object ClickRankingRetry : HomeSideEffect()
 }
